Add Recent Chat API

+++ FamilyChatAPI/FamilyChatAPI/Dtos/ChatHubDto/PersonalMessageDto.cs
@@ -5,9 +5,9 @@
         public string connectionId { get; set; }
         public bool isUser { get; set; }
         public string passWord { get; set; }
-        public int userId { get; set; }
+        public byte userId { get; set; }
         public string userName { get; set; }
-        public int chatId { get; set; }
+        public byte chatId { get; set; }
         public string messageText { get; set; }
     }
 }
